Add files via upload
Sample Database: hospital_management_system

With the help of a Hospital Database, this exercises will help you understand simple SQL select queries to advanced multi table JOIN queries.

Sample Database description:

Hospitals are the most important part of our lives, trying to provide the best medical facilities to people suffering from various type of illness, which may be due to change in climate conditions, increased work-load, emotional trauma stress etc. It is very much difficult for the hospital to maintain its day-to-day activities and records manually. That is why a database is required to keep records of all type of activities of a hospital.


List of Tables:
*Patient
*Doctor
*lab
*in_patient
*out_patient
*bill
*room